Move voucher-related forms, signals and views to the voucher application.
