+++ src/oscar/apps/basket/signals.py
@@ -2,7 +2,3 @@ import django.dispatch
 
 basket_addition = django.dispatch.Signal(
     providing_args=["product", "user", "request"])
-voucher_addition = django.dispatch.Signal(
-    providing_args=["basket", "voucher"])
-voucher_removal = django.dispatch.Signal(
-    providing_args=["basket", "voucher"])
