+++ src/oscar/apps/voucher/apps.py
@@ -1,6 +1,8 @@
+from django.urls import path
 from django.utils.translation import gettext_lazy as _
 
 from oscar.core.application import OscarConfig
+from oscar.core.loading import get_class
 
 
 class VoucherConfig(OscarConfig):
@@ -11,3 +13,13 @@ class VoucherConfig(OscarConfig):
     def ready(self):
         from . import receivers  # noqa
         from . import signals  # noqa
+
+        self.add_voucher_view = get_class('voucher.views', 'VoucherAddView')
+        self.remove_voucher_view = get_class('voucher.views', 'VoucherRemoveView')
+
+    def get_urls(self):
+        urls = [
+            path('vouchers/add/', self.add_voucher_view.as_view(), name='vouchers-add'),
+            path('vouchers/<int:pk>/remove/', self.remove_voucher_view.as_view(), name='vouchers-remove'),
+        ]
+        return self.post_process_urls(urls)
